Add tests for filtering based on tags
