+++ src/main.rs
@@ -45,13 +45,6 @@ struct Opts {
     #[options(no_short, help = "Export only files with this tag")]
     only_tags: Vec<String>,
 
-    #[options(
-        no_short,
-        help = "Exclude files with this flag in the frontmatter from the export",
-        default = "private"
-    )]
-    ignore_frontmatter_flag: String,
-
     #[options(no_short, help = "Export hidden files", default = "false")]
     hidden: bool,
 
@@ -68,14 +61,6 @@ struct Opts {
     )]
     hard_linebreaks: bool,
 }
-
-// fn comma_separated(input: &str) -> Result<Vec<String>> {
-//     Ok(if input.is_empty() {
-//         Vec::new()
-//     } else {
-//         input.split(",").map(|s| s.trim().to_string()).collect()
-//     })
-// }
 
 fn frontmatter_strategy_from_str(input: &str) -> Result<FrontmatterStrategy> {
     match input {
@@ -114,9 +99,6 @@ fn main() {
     if args.hard_linebreaks {
         exporter.add_postprocessor(&softbreaks_to_hardbreaks);
     }
-
-    let frontmatter_flag_postprocessor = filter_by_frontmatter_flag(args.ignore_frontmatter_flag);
-    exporter.add_postprocessor(&frontmatter_flag_postprocessor);
 
     let tags_postprocessor = filter_by_tags(args.skip_tags, args.only_tags);
     exporter.add_postprocessor(&tags_postprocessor);
