+++ tests/postprocessors_test.rs
@@ -1,4 +1,4 @@
-use obsidian_export::postprocessors::softbreaks_to_hardbreaks;
+use obsidian_export::postprocessors::{filter_by_tags, softbreaks_to_hardbreaks};
 use obsidian_export::{Context, Exporter, MarkdownEvents, PostprocessorResult};
 use pretty_assertions::assert_eq;
 use pulldown_cmark::{CowStr, Event};
@@ -8,6 +8,7 @@ use std::fs::{read_to_string, remove_file};
 use std::path::PathBuf;
 use std::sync::Mutex;
 use tempfile::TempDir;
+use walkdir::WalkDir;
 
 /// This postprocessor replaces any instance of "foo" with "bar" in the note body.
 fn foo_to_bar(_ctx: &mut Context, events: &mut MarkdownEvents) -> PostprocessorResult {
@@ -246,4 +247,46 @@ fn test_softbreaks_to_hardbreaks() {
         read_to_string("tests/testdata/expected/postprocessors/hard_linebreaks.md").unwrap();
     let actual = read_to_string(tmp_dir.path().join(PathBuf::from("hard_linebreaks.md"))).unwrap();
     assert_eq!(expected, actual);
+}
+
+#[test]
+fn test_filter_by_tags() {
+    let tmp_dir = TempDir::new().expect("failed to make tempdir");
+    let mut exporter = Exporter::new(
+        PathBuf::from("tests/testdata/input/filter-by-tags"),
+        tmp_dir.path().to_path_buf(),
+    );
+    let filter_by_tags = filter_by_tags(
+        vec!["private".to_string(), "no-export".to_string()],
+        vec!["export".to_string()],
+    );
+    exporter.add_postprocessor(&filter_by_tags);
+    exporter.run().unwrap();
+
+    let walker = WalkDir::new("tests/testdata/expected/filter-by-tags/")
+        // Without sorting here, different test runs may trigger the first assertion failure in
+        // unpredictable order.
+        .sort_by(|a, b| a.file_name().cmp(b.file_name()))
+        .into_iter();
+    for entry in walker {
+        let entry = entry.unwrap();
+        if entry.metadata().unwrap().is_dir() {
+            continue;
+        };
+        let filename = entry.file_name().to_string_lossy().into_owned();
+        let expected = read_to_string(entry.path()).unwrap_or_else(|_| {
+            panic!(
+                "failed to read {} from testdata/expected/filter-by-tags",
+                entry.path().display()
+            )
+        });
+        let actual = read_to_string(tmp_dir.path().join(PathBuf::from(&filename)))
+            .unwrap_or_else(|_| panic!("failed to read {} from temporary exportdir", filename));
+
+        assert_eq!(
+            expected, actual,
+            "{} does not have expected content",
+            filename
+        );
+    }
 }
