+++ tests/testdata/expected/filter-by-tags/export-me.md
@@ -1,0 +1,7 @@
+---
+tags:
+- export
+- me
+---
+
+A public note
