+++ tests/testdata/expected/filter-by-tags/export.md
@@ -1,0 +1,6 @@
+---
+tags:
+- export
+---
+
+A public note
