--- tests/testdata/expected/ignore-default-keyword/do-expört.md
@@ -1,5 +1,0 @@
----
-no-expört: false
----
-
-A note with negated special ignore keyword
