--- tests/testdata/expected/ignore-default-keyword/no-expört.md
@@ -1,5 +1,0 @@
----
-no-expört: true
----
-
-A note with a special ignore keyword
