--- tests/testdata/expected/ignore-specific-keyword/do-expört.md
@@ -1,5 +1,0 @@
----
-no-expört: false
----
-
-A note with negated special ignore keyword
