--- tests/testdata/expected/ignore-specific-keyword/private.md
@@ -1,5 +1,0 @@
----
-private: true
----
-
-A private note.
