+++ tests/testdata/input/filter-by-tags/export-me.md
@@ -1,0 +1,5 @@
+---
+tags: [export, me]
+---
+
+A public note
