+++ tests/testdata/input/filter-by-tags/export-no-export.md
@@ -1,0 +1,5 @@
+---
+tags: [export, no-export, private]
+---
+
+A private note
