+++ tests/testdata/input/filter-by-tags/export.md
@@ -1,0 +1,5 @@
+---
+tags: [export]
+---
+
+A public note
