+++ tests/testdata/input/filter-by-tags/no-no-export.md
@@ -1,0 +1,5 @@
+---
+tags: [no, no-export]
+---
+
+A private note
