+++ tests/testdata/input/filter-by-tags/no-tags.md
@@ -1,5 +1,5 @@
 ---
-private: false
+title: foo
 ---
 
 A public note.
