--- tests/testdata/input/ignore-keyword/do-expört.md
@@ -1,4 +1,0 @@
----
-no-expört: false
----
-A note with negated special ignore keyword
