--- tests/testdata/input/ignore-keyword/no-expört.md
@@ -1,4 +1,0 @@
----
-no-expört: true
----
-A note with a special ignore keyword
