--- tests/testdata/input/ignore-keyword/no-frontmatter.md
@@ -1,1 +1,0 @@
-A note without frontmatter should be exported.
