--- tests/testdata/input/ignore-keyword/public.md
@@ -1,4 +1,0 @@
----
-private: false
----
-A public note.
